Refactor Contact page to implement a new form layout and remove old 3D component
